requirements.txt
opencv-python
Flask
python-dotenv

tqdm==4.65.0
pandas==1.4.4
PyMySQL==1.0.2
SQLAlchemy==1.4.41